requirements.txt
build==0.10.0
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
packaging==23.1
pyproject_hooks==1.0.0
python-dotenv==1.0.0
requests==2.29.0
requests-mock==1.10.0
six==1.16.0
urllib3==1.26.15
